organize files in app/src/assets directory
