+++ app/src/assets/exercise-cards/bridge.png
