+++ app/src/assets/exercise-cards/deadBug.png
