+++ app/src/assets/exercise-cards/latExtRotation.png
