+++ app/src/assets/exercise-cards/muscleUp.png
