+++ app/src/assets/exercise-cards/pullUp.png
