+++ app/src/assets/exercise-cards/pushUp.png
