+++ app/src/assets/exercise-cards/squat.png
