+++ app/src/assets/instructions/squatHelp.png
