+++ app/src/assets/logos/productNameBlack.png
