+++ app/src/assets/logos/productNameWhite.png
