+++ app/src/pages/Catalog.js
@@ -2,13 +2,13 @@ import React from "react";
 import { Typography, Box } from "@mui/material";
 import Grid from "@mui/material/Grid2";
 import ExerciseCard from "../components/ExerciseCard.js";
-import squatImg from "../assets/squat.png";
-import bridgeImg from "../assets/bridge.png";
-import deadBugImage from "../assets/deadBug.png";
-import pushUpImage from "../assets/pushUp.png";
-import pullUpImage from "../assets/pullUp.png";
-import muscleUpImage from "../assets/muscleUp.png";
-import latExtRotationImg from "../assets/latExtRotation.png";
+import squatImg from "../assets/exercise-cards/squat.png";
+import bridgeImg from "../assets/exercise-cards/bridge.png";
+import deadBugImage from "../assets/exercise-cards/deadBug.png";
+import pushUpImage from "../assets/exercise-cards/pushUp.png";
+import pullUpImage from "../assets/exercise-cards/pullUp.png";
+import muscleUpImage from "../assets/exercise-cards/muscleUp.png";
+import latExtRotationImg from "../assets/exercise-cards/latExtRotation.png";
 import { catalogText } from "../assets/content.js";
 
 /**
